Anomaly uses existential quantification
The problem: we want Anomaly to carry the edge label and vertex relevant
to the anomaly (edges with given label violate cardinality constraint at
given vertex) but we want multi-hop walks in a graph to produce Either
Anomaly x rather than Either Anomaly (Either Anomaly ( ... (Either
Anomaly x) ... )) which would be necessary if we used universal
quantification because, well, what would be the type of the multi-hop
walk?

  Either (Anomaly ? (Vertex m)) (m a)

With existential quantification, we can just throw any edge label onto
an anomaly, and all the user can do is show it.